[CLI] Include Memory Scheduling validation checks for Compute Resources with multiple instance types
- Memory Scheduling Requirements:
    - Memory-based scheduling is supported for Compute Resource that use either 'InstanceType' or have a single instance type under 'InstanceTypeList'
